+++ src/utils/normalize-path.ts
@@ -1,0 +1,5 @@
+export const normalizePath = (filePath: string) => (
+	/^[./]/.test(filePath)
+		? filePath
+		: `./${filePath}`
+);
